[ApiBundle][PromotionCoupon] Flattening structure for coupon generation
