+++ src/Sylius/Behat/Context/Api/Admin/ManagingPromotionCouponsContext.php
@@ -144,7 +144,7 @@ final class ManagingPromotionCouponsContext implements Context
      */
     public function iSpecifyItsAmountAs(int $amount): void
     {
-        $this->request->updateContent(['instruction' => ['amount' => $amount]]);
+        $this->request->updateContent(['amount' => $amount]);
     }
 
     /**
@@ -152,7 +152,7 @@ final class ManagingPromotionCouponsContext implements Context
      */
     public function specifyPrefixAs(string $prefix): void
     {
-        $this->request->updateContent(['instruction' => ['prefix' => $prefix]]);
+        $this->request->updateContent(['prefix' => $prefix]);
     }
 
     /**
@@ -160,7 +160,7 @@ final class ManagingPromotionCouponsContext implements Context
      */
     public function specifySuffixAs(string $suffix): void
     {
-        $this->request->updateContent(['instruction' => ['suffix' => $suffix]]);
+        $this->request->updateContent(['suffix' => $suffix]);
     }
 
     /**
@@ -169,7 +169,7 @@ final class ManagingPromotionCouponsContext implements Context
      */
     public function iSpecifyTheirCodeLengthAs(?int $codeLength = null): void
     {
-        $this->request->updateContent(['instruction' => ['codeLength' => $codeLength]]);
+        $this->request->updateContent(['codeLength' => $codeLength]);
     }
 
     /**
@@ -177,7 +177,7 @@ final class ManagingPromotionCouponsContext implements Context
      */
     public function iSetGeneratedCouponsUsageLimitTo(int $limit):void
     {
-        $this->request->updateContent(['instruction' => ['usageLimit' => $limit]]);
+        $this->request->updateContent(['usageLimit' => $limit]);
     }
 
     /**
@@ -185,7 +185,7 @@ final class ManagingPromotionCouponsContext implements Context
      */
     public function iMakeGeneratedCouponsValidUntil(\DateTimeInterface $date): void
     {
-        $this->request->updateContent(['instruction' => ['expiresAt' => $date->format('Y-m-d')]]);
+        $this->request->updateContent(['expiresAt' => $date->format('Y-m-d')]);
     }
 
     /**
@@ -205,10 +205,9 @@ final class ManagingPromotionCouponsContext implements Context
     }
 
     /**
-     * @When I (try to) generate it
      * @When I (try to) generate these coupons
      */
-    public function iResetIt(): void
+    public function iGenerateTheseCoupons(): void
     {
         $this->client->executeCustomRequest($this->request);
     }
@@ -548,7 +547,7 @@ final class ManagingPromotionCouponsContext implements Context
     {
         Assert::same(
             $this->responseChecker->getError($this->client->getLastResponse()),
-            'instruction.amount: Please enter amount of coupons to generate.',
+            'amount: Please enter amount of coupons to generate.',
         );
     }
 
@@ -559,7 +558,7 @@ final class ManagingPromotionCouponsContext implements Context
     {
         Assert::same(
             $this->responseChecker->getError($this->client->getLastResponse()),
-            'instruction.codeLength: Please enter coupon code length.',
+            'codeLength: Please enter coupon code length.',
         );
     }
 
@@ -570,7 +569,7 @@ final class ManagingPromotionCouponsContext implements Context
     {
         Assert::same(
             $this->responseChecker->getError($this->client->getLastResponse()),
-            'instruction.codeLength: Coupon code length must be between 1 and 40.',
+            'codeLength: Coupon code length must be between 1 and 40.',
         );
     }
