+++ src/Sylius/Bundle/ApiBundle/Command/Promotion/GeneratePromotionCoupon.php
@@ -13,12 +13,19 @@ declare(strict_types=1);
 
 namespace Sylius\Bundle\ApiBundle\Command\Promotion;
 
+use Sylius\Component\Promotion\Generator\PromotionCouponGeneratorInstructionInterface;
+
 /** @experimental */
-class GeneratePromotionCoupon
+class GeneratePromotionCoupon implements PromotionCouponGeneratorInstructionInterface
 {
     public function __construct(
         private string $promotionCode,
-        private PromotionCouponGeneratorInstruction $instruction,
+        private ?string $prefix = null,
+        private ?int $codeLength = null,
+        private ?string $suffix = null,
+        private ?int $amount = null,
+        private ?\DateTimeInterface $expiresAt = null,
+        private ?int $usageLimit = null,
     ) {
     }
 
@@ -27,8 +34,63 @@ class GeneratePromotionCoupon
         return $this->promotionCode;
     }
 
-    public function getInstruction(): PromotionCouponGeneratorInstruction
+    public function getAmount(): int
     {
-        return $this->instruction;
+        return $this->amount;
+    }
+
+    public function setAmount(?int $amount): void
+    {
+        $this->amount = $amount;
+    }
+
+    public function getCodeLength(): int
+    {
+        return $this->codeLength;
+    }
+
+    public function setCodeLength(?int $codeLength): void
+    {
+        $this->codeLength = $codeLength;
+    }
+
+    public function getPrefix(): ?string
+    {
+        return $this->prefix;
+    }
+
+    public function setPrefix(?string $prefix): void
+    {
+        $this->prefix = $prefix;
+    }
+
+    public function getSuffix(): ?string
+    {
+        return $this->suffix;
+    }
+
+    public function setSuffix(?string $suffix): void
+    {
+        $this->suffix = $suffix;
+    }
+
+    public function getExpiresAt(): ?\DateTimeInterface
+    {
+        return $this->expiresAt;
+    }
+
+    public function setExpiresAt(?\DateTimeInterface $expiresAt): void
+    {
+        $this->expiresAt = $expiresAt;
+    }
+
+    public function getUsageLimit(): ?int
+    {
+        return $this->usageLimit;
+    }
+
+    public function setUsageLimit(?int $usageLimit): void
+    {
+        $this->usageLimit = $usageLimit;
     }
 }
