+++ src/Sylius/Bundle/ApiBundle/Resources/config/serialization/Commands/Promotion/GeneratePromotionCoupon.xml
@@ -19,7 +19,22 @@
         <attribute name="promotionCode">
             <group>admin:promotion_coupon:update</group>
         </attribute>
-        <attribute name="instruction">
+        <attribute name="prefix">
+            <group>admin:promotion_coupon:update</group>
+        </attribute>
+        <attribute name="codeLength">
+            <group>admin:promotion_coupon:update</group>
+        </attribute>
+        <attribute name="suffix">
+            <group>admin:promotion_coupon:update</group>
+        </attribute>
+        <attribute name="amount">
+            <group>admin:promotion_coupon:update</group>
+        </attribute>
+        <attribute name="expiresAt">
+            <group>admin:promotion_coupon:update</group>
+        </attribute>
+        <attribute name="usageLimit">
             <group>admin:promotion_coupon:update</group>
         </attribute>
     </class>
