--- src/Sylius/Bundle/ApiBundle/Resources/config/serialization/Commands/Promotion/PromotionCouponGeneratorInstruction.xml
@@ -1,38 +1,0 @@
-<?xml version="1.0" ?>
-
-<!--
-
- This file is part of the Sylius package.
-
- (c) Sylius Sp. z o.o.
-
- For the full copyright and license information, please view the LICENSE
- file that was distributed with this source code.
-
--->
-
-<serializer xmlns="http://symfony.com/schema/dic/serializer-mapping"
-            xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-            xsi:schemaLocation="http://symfony.com/schema/dic/serializer-mapping https://symfony.com/schema/dic/serializer-mapping/serializer-mapping-1.0.xsd"
->
-    <class name="Sylius\Bundle\ApiBundle\Command\Promotion\PromotionCouponGeneratorInstruction">
-        <attribute name="prefix">
-            <group>admin:promotion_coupon:update</group>
-        </attribute>
-        <attribute name="codeLength">
-            <group>admin:promotion_coupon:update</group>
-        </attribute>
-        <attribute name="suffix">
-            <group>admin:promotion_coupon:update</group>
-        </attribute>
-        <attribute name="amount">
-            <group>admin:promotion_coupon:update</group>
-        </attribute>
-        <attribute name="expiresAt">
-            <group>admin:promotion_coupon:update</group>
-        </attribute>
-        <attribute name="usageLimit">
-            <group>admin:promotion_coupon:update</group>
-        </attribute>
-    </class>
-</serializer>
