+++ src/Sylius/Bundle/ApiBundle/Resources/config/validation/GeneratePromotionCoupon.xml
@@ -20,8 +20,45 @@
                 </option>
             </constraint>
         </property>
-        <property name="instruction">
-            <constraint name="Valid" />
+        <property name="amount">
+            <constraint name="NotBlank">
+                <option name="message">sylius.promotion_coupon_generator_instruction.amount.not_blank</option>
+                <option name="groups">
+                    <value>sylius</value>
+                </option>
+            </constraint>
+            <constraint name="Range">
+                <option name="min">1</option>
+                <option name="minMessage">sylius.promotion_coupon_generator_instruction.amount.min</option>
+                <option name="groups">
+                    <value>sylius</value>
+                </option>
+            </constraint>
+        </property>
+        <property name="codeLength">
+            <constraint name="NotBlank">
+                <option name="message">sylius.promotion_coupon_generator_instruction.code_length.not_blank</option>
+                <option name="groups">
+                    <value>sylius</value>
+                </option>
+            </constraint>
+            <constraint name="Range">
+                <option name="min">1</option>
+                <option name="max">40</option>
+                <option name="notInRangeMessage">sylius.promotion_coupon_generator_instruction.code_length.not_in_range</option>
+                <option name="groups">
+                    <value>sylius</value>
+                </option>
+            </constraint>
+        </property>
+        <property name="usageLimit">
+            <constraint name="Range">
+                <option name="min">1</option>
+                <option name="minMessage">sylius.promotion_coupon_generator_instruction.usage_limit.min</option>
+                <option name="groups">
+                    <value>sylius</value>
+                </option>
+            </constraint>
         </property>
     </class>
 </constraint-mapping>
